more

+++ include/builders/util.hpp
@@ -77,7 +77,7 @@ struct build_configuration {
     bool verbose_output;
 };
 
-static uint64_t compute_num_buckets(const uint64_t num_keys, const double avg_bucket_size) {
+static uint64_t compute_num_buckets(const uint64_t num_keys, const uint64_t avg_bucket_size) {
     return std::ceil(static_cast<double>(num_keys) / avg_bucket_size);
 }
